move db to common
